rrr: convert base/threading to module (rrr.threading)
threading.cpp becomes rrr.threading module interface unit; threading.hpp
deleted. ~1000-line merged file: Pthread_* wrappers, Lockable, SpinLock,
SpinMutex<T> family (SpinMutexGuard, SpinPoisonError, SpinLockResult),
SpinCondVar, Queue<T>, ThreadPool, RunLater.

Includers (9) updated to `import rrr.threading;` placed after textual
#includes (libc++ ordering constraint):
  - src/rrr/rpc/{fiber_channel,tcp_channel,inmemory_channel,callbacks,
    request_queue}.hpp
  - src/rrr/misc/{marshal,serializable,any_message}.cpp
  - src/rrr/base/all.hpp, src/rrr/rrr.hpp

One small fix from the `using namespace std;` removal: `max(0.0, ...)`
in RunLater::max_wait → `std::max(0.0, ...)`.

Measurement (rrr + rpcbench, -j32, clean):
  +misc:       30.14 s, 112.4 MB PCM
  +threading:  37.01 s, 136.8 MB PCM, librrr.a 9.99 MB

+6.9 s wallclock (threading is heavily fanned-out: 9 includers, all in
the hot RPC/reactor path). 2.19× baseline; abort threshold (3×)
remains ~50 s.
