+++ src/rrr/base/all.hpp
@@ -3,8 +3,6 @@
 //#include "unittest.hpp"
 
 
-#include "threading.hpp"
-
 // Imports go after textual #includes (libc++ rejects `import std; ...
 // #include <vector>` order).
 import rrr.basetypes;
@@ -12,3 +10,4 @@ import rrr.debugging;
 import rrr.logging;
 import rrr.misc;
 import rrr.strop;
+import rrr.threading;
