+++ src/rrr/rpc/tcp_channel.hpp
@@ -14,11 +14,12 @@
 
 
 
-#include "../base/threading.hpp"
 #include "channel.hpp"
 #include "frame_codec.hpp"
 #include "pollable_proxy.h"
 #include "../reactor/reactor.h"
+
+import rrr.threading;
 
 /**
  * SRPC TCP Channel Backend.
